+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/op/OpPipelineCallbackResource.kt
@@ -1,10 +1,20 @@
 package com.tencent.devops.process.api.op
 
+import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_USER_ID
 import com.tencent.devops.common.api.pojo.Result
+import com.tencent.devops.common.pipeline.event.CallBackEvent
+import com.tencent.devops.common.pipeline.event.CallBackNetWorkRegionType
+import com.tencent.devops.common.pipeline.event.ProjectPipelineCallBack
+import com.tencent.devops.common.pipeline.pojo.secret.ISecretParam
+import com.tencent.devops.process.pojo.CreateCallBackResult
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
 import javax.ws.rs.Consumes
+import javax.ws.rs.DELETE
+import javax.ws.rs.GET
+import javax.ws.rs.HeaderParam
+import javax.ws.rs.POST
 import javax.ws.rs.PUT
 import javax.ws.rs.Path
 import javax.ws.rs.PathParam
@@ -53,4 +63,49 @@ interface OpPipelineCallbackResource {
         @QueryParam("callbackIds")
         callbackIds: String
     ): Result<Boolean>
+
+
+    @POST
+    @Path("/create")
+    @Operation(summary = "新建项目级回调")
+    fun createProjectCallback(
+        @Parameter(description = "userId", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_USER_ID)
+        userId: String,
+        @Parameter(description = "事件类型", required = true)
+        @QueryParam("event")
+        eventType: CallBackEvent,
+        @Parameter(description = "回调地址", required = true)
+        @QueryParam("url")
+        url: String,
+        @Parameter(description = "region", required = false)
+        @QueryParam("region")
+        region: CallBackNetWorkRegionType?,
+        @Parameter(description = "回调凭证", required = true)
+        secretParam: ISecretParam
+    ): Result<CreateCallBackResult>
+
+    @DELETE
+    @Path("/delete/{id}")
+    @Operation(summary = "移除回调")
+    fun delete(
+        @Parameter(description = "userId", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_USER_ID)
+        userId: String,
+        @PathParam("id")
+        @Parameter(description = "事件ID", required = true)
+        id: Long
+    ): Result<Boolean>
+
+    @GET
+    @Path("/list")
+    @Operation(summary = "查询项目级回调")
+    fun list(
+        @Parameter(description = "userId", required = true)
+        @HeaderParam(AUTH_HEADER_DEVOPS_USER_ID)
+        userId: String,
+        @QueryParam("event")
+        @Parameter(description = "事件类型", required = true)
+        event: CallBackEvent
+    ): Result<List<ProjectPipelineCallBack>>
 }
